added view file
This has an error on it. The error states that it has an operator that does not exist and the characters are varying from text to integer. Since I cannot compare an integer with a varchar. I asked advice to my classmates how should I approach this error, and they were also stumped when I showed them my code. I decided to ask the internet and it showed to use " : : " as the casting syntax.

Yet the problem is, it is not the tables that are needing a casting syntax, but the " TRUNC(AVG((vbm_total_vaccinations/365))) " since it returns an integer.